fibonacci from long to long long

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
 int main(void)
 {
 	int count;
-	long int f1 = 1;
-	long int f2 = 2;
-	long int next;
+	long long int f1 = 1;
+	long long int f2 = 2;
+	long long int next;
 	int flimit = 50;
 
 	for (count = 2; count <= flimit; count++)
@@ -18,7 +18,7 @@ int main(void)
 		next = f1 + f2;
 		f1 = f2;
 		f2 = next;
-		printf("%lu", next);
+		printf("%llu", next);
 		if (count < flimit)
 			printf(", ");
 	}
